Don't try to guess the hook url. Allow updating all repository settings.
Previously, we made the assumption that the redmine/chiliproject installation was being served on the HTTP port. This is no true for ALL installations, it can also be served       through HTTPS. So, now, we store the rails generated URL and pass that to the settings. Instead of adding another config variable to all repositories(https_server, for example),   we renamed `server` to `url` which suites our needs better.
We now also provide a checkbox to allow updating ALL repositories settings, ie, force update settings.
Separated code from html on the project repository settings patch.