Add user-edit screen for public_keys and fix security flaw.
This commit refactors the gitolite-public-key view to (1) remove
MVC violations (i.e. move the DB access into respective controllers,
(2) separate the view/edit functionality into a separate partial so
that (3) it can be called from the user edit screen.  As a result,
the administrator can now look at and edit user's keys (which I have
found is necessary for some clueless users).  The previous commit
would allow someone to view other public keys (but not edit them!); this
commit fixes this flaw.

Also, updated the style of the display to better reflect the 1.3+
style of redmine.  This code is still backward compatible with
earlier versions of redmine.  Also, moved .rhtml => .html.erb in
preparation for Rails 3.0 (rest of Redmine alreday updated)