Properly handle the archiving of projects.
When a project is archived, we remove its keys from gitolite.conf (and
possibly from the keydir if this project is the only one that uses
these keys).  We also remove the git_daemon access and mark the gitolite.conf
entry with a dummy key: "redmine_archived_project".

We leave the repository in the gitolite repo.