First take on implementation of deployment credentials.
Deployment credentials are constructed on the repository settings page
for a repository and consist of combinations of public keys and
permissions (currently either "R" or "RW+".  The public keys for
deployment credentials are owned by users with 'manager' permissions.
The administrator can modify (add/update/delete) all deployment
credentials.

Also included with this commit is a functioning validation on the
public key display window (this wasn't working originally).