Add execution of additional post-receive hooks
As this plugin occupies the post-receive hook of a managed repository,
no other post-receive hook can be used. To get around this limitation,
this commit changes the installed hook in such a way that it looks for
additional executable files in a sub-directory of the repository's hook
folder. This sub-directory has to be called "post-receive.d". The
plugin's hook will execute all the files in there marked as executable
and will provide the processes with the same STDIN stream it got when
called in the first place.