Correct loading of RedmineGitHosting::Commands & Config modules
Maybe due to upgrading to Ruby 3.x, the code of these modules couldn't
be loaded anymore, due to an infinite loading loop.

For instance, for the Commands module:
* Loading of commands.rb is triggered,
  * As its first instruction, the module extends Commands::Base,
  * So the loading of commands/base.rb is triggered,
    * But then, in that file, the Base module is included in the
      Commands module, which has to be loaded first,
      * And so, we're back to the beginning...

To work around this, the "extend Xx::Xx" instructions are moved to the
submodules themselves, so that each submodule can be loaded
independently and add its contributions to the main module (Commands or
Config).